+++ packages/react-native-sdk/__tests__/components/CallParticipantsView.test.tsx
@@ -5,7 +5,7 @@ import mockParticipant from '../mocks/participant';
 import { A11yComponents } from '../../src/constants/A11yLabels';
 import { mockCall } from '../mocks/call';
 import { act, render, screen, within } from '../utils/RNTLTools';
-import { CallParticipantsView } from '../../src/components';
+import { CallContentView } from '../../src/components';
 import { ViewToken } from 'react-native';
 
 console.warn = jest.fn();
@@ -50,7 +50,7 @@ describe('CallParticipantsView', () => {
       }),
     ]);
 
-    render(<CallParticipantsView />, {
+    render(<CallContentView />, {
       call,
     });
 
@@ -100,7 +100,7 @@ describe('CallParticipantsView', () => {
       }),
     ]);
 
-    render(<CallParticipantsView />, {
+    render(<CallContentView />, {
       call,
     });
 
@@ -143,7 +143,7 @@ describe('CallParticipantsView', () => {
       }),
     ]);
 
-    render(<CallParticipantsView />, {
+    render(<CallContentView />, {
       call,
     });