+++ sample-apps/react-native/dogfood/src/components/ActiveCall.tsx
@@ -1,21 +1,25 @@
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import {
+  CallContentViewProps,
   CallControlsView,
   CallControlsViewType,
-  CallParticipantsBadge,
-  CallParticipantsView,
   CallingState,
+  ParticipantsInfoBadge,
+  CallContentView,
   useCall,
   useIncallManager,
 } from '@stream-io/video-react-native-sdk';
-import { ActivityIndicator, StyleSheet } from 'react-native';
+import { ActivityIndicator, StyleSheet, View } from 'react-native';
 import { appTheme } from '../theme';
 import {
   SafeAreaView,
   useSafeAreaInsets,
 } from 'react-native-safe-area-context';
+import { ParticipantsLayoutButtons } from './ParticipantLayoutButton';
 
 type ActiveCallProps = CallControlsViewType;
+
+type Layout = CallContentViewProps['mode'];
 
 export const ActiveCall = ({
   chatButton,
@@ -24,6 +28,7 @@ export const ActiveCall = ({
   const call = useCall();
   const activeCallRef = useRef(call);
   activeCallRef.current = call;
+  const [selectedLayout, setSelectedLayout] = useState<Layout>('grid');
 
   useEffect(() => {
     return () => {
@@ -46,8 +51,14 @@ export const ActiveCall = ({
 
   return (
     <SafeAreaView style={styles.container} edges={['top', 'left', 'right']}>
-      <CallParticipantsBadge style={[styles.iconGroup, { top }]} />
-      <CallParticipantsView />
+      <View style={[styles.icons, { top }]}>
+        <ParticipantsLayoutButtons
+          selectedLayout={selectedLayout}
+          setSelectedLayout={setSelectedLayout}
+        />
+        <ParticipantsInfoBadge />
+      </View>
+      <CallContentView mode={selectedLayout} />
       <CallControlsView
         chatButton={chatButton}
         hangUpCallButton={hangUpCallButton}
@@ -65,10 +76,12 @@ const styles = StyleSheet.create({
     flex: 1,
     backgroundColor: appTheme.colors.static_grey,
   },
-  iconGroup: {
+  icons: {
     position: 'absolute',
-    marginTop: appTheme.spacing.lg,
-    right: appTheme.spacing.sm,
+    right: 0,
+    marginTop: appTheme.spacing.sm,
+    flexDirection: 'row',
+    alignItems: 'center',
     zIndex: appTheme.zIndex.IN_FRONT,
   },
   callControlsWrapper: {