+++ sample-apps/react/react-video-demo/src/hooks/useCallStateNotification.tsx
@@ -1,0 +1,30 @@
+import { useEffect } from 'react';
+import { v1 as uuid } from 'uuid';
+import { CallingState, useCallCallingState } from '@stream-io/video-react-sdk';
+import { Info } from '../components/Icons';
+
+import { useNotificationContext } from '../contexts/NotificationsContext';
+
+export const useCallStateNotification = () => {
+  const callingState = useCallCallingState();
+
+  const { addNotification } = useNotificationContext();
+
+  useEffect(() => {
+    if (callingState === CallingState.OFFLINE) {
+      addNotification({
+        id: uuid(),
+        message: 'No internet connection',
+        icon: <Info />,
+      });
+    }
+
+    if (callingState === CallingState.RECONNECTING) {
+      addNotification({
+        id: uuid(),
+        message: 'Reconnecting...',
+        icon: <Info />,
+      });
+    }
+  }, [callingState, addNotification]);
+};